# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: SCM Procurement Mock Server API
  version: 1.0.0
  description: |
    Detailed API specification for the mock SCM procurement backend.

    Notes:
    - Most business endpoints expect JWT auth via `Authorization: Bearer <token>`.
    - `/chat/realtime/ws` and `/ws/{user_id}` are WebSocket endpoints and are documented here as reference.
servers:
  - url: http://localhost:8000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    IntegrationStartChatThreadRequest:
      type: object
      required: [fromEmail, fromName, toEmail, toName]
      properties:
        fromEmail:
          type: string
          format: email
        fromName:
          type: string
        toEmail:
          type: string
          format: email
        toName:
          type: string
    IntegrationAddChatParticipantRequest:
      type: object
      required: [email, name, fromEmail]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        fromEmail:
          type: string
          format: email
...
paths:
...
  /chat:
    post:
      tags: [Integration Chat]
      summary: Start integration chat thread
      operationId: integrationStartChatThread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationStartChatThreadRequest'
      responses:
        '201':
          description: Chat thread started or reused
          content:
            application/json:
              schema:
                type: object
  /chat/add:
    patch:
      tags: [Integration Chat]
      summary: Add participant to integration chat thread
      operationId: integrationAddParticipant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationAddChatParticipantRequest'
      responses:
        '200':
          description: Participant added
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Chat thread not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...